test: relocate reviewer-fix regressions to their home suites, drop duplicates
A prior partial move (b15fe5365) scattered copies of these tests into module-specific files but left
test_reviewer_fixes.py in place with the originals — so several classes existed in two files at once
(CodeRabbit flagged TestSingleCableVerifyServerKey as a verbatim duplicate). Finish the job so each of
the 8 classes lives in exactly one home:

- normalize-mapping + create-platform-full-clean → test_coverage_device_fields
- generate/load VC member name → test_virtual_chassis
- all-server-mappings validation → test_coverage_sync_view (beside TestBuildAllServerMappings)
- render-device-selection escape → test_coverage_tables
- Drop the two verbatim duplicates (SingleCableVerifyServerKey, ImportSingleDeviceLazyValidation) —
  canonical copies already live in test_view_wiring / test_import_utils; also remove the misplaced
  copies the prior move had dumped into test_view_wiring.
- Delete test_reviewer_fixes.py.

De-mock where the real object drives the behaviour: the render-escape test builds a real virtual
chassis + a real member Device with an XSS name; the VC-member-name loader reads a real LibreNMSSettings
row (resolved via apps.get_model and restored over the session-wide LibreNMSSettings mock for the
loader's deferred import). None/boolean/DB-error cases keep a mock — a NOT-NULL CharField cannot hold
those values. Full suite green (3337).
