+++ netbox_librenms_plugin/tests/test_virtual_chassis.py
@@ -2,59 +2,85 @@
 
 from unittest.mock import MagicMock, patch
 
+import pytest
 
+
+@pytest.mark.django_db
 class TestLoadVcMemberNamePattern:
     """_load_vc_member_name_pattern must return valid string or default."""
 
     DEFAULT = "-M{position}"
+
+    @staticmethod
+    def _real_settings_model():
+        # apps.get_model returns the REAL registered model even when the session-wide
+        # mock_librenms_config autouse has patched netbox_librenms_plugin.models.LibreNMSSettings to a
+        # MagicMock — so real rows can be persisted and read back.
+        from django.apps import apps
+
+        return apps.get_model("netbox_librenms_plugin", "LibreNMSSettings")
 
     def _call(self):
         from netbox_librenms_plugin.import_utils.virtual_chassis import _load_vc_member_name_pattern
 
         return _load_vc_member_name_pattern()
 
+    def _call_real(self):
+        # The loader does a deferred `from ...models import LibreNMSSettings`, which reads the autouse
+        # mock in the full suite. Restore the real model for the duration of the call so it queries the
+        # row _store_pattern persisted (the whole point of the real-DB de-mock).
+        with patch("netbox_librenms_plugin.models.LibreNMSSettings", self._real_settings_model()):
+            return self._call()
+
+    def _store_pattern(self, pattern):
+        """Persist a single real LibreNMSSettings row carrying the given pattern."""
+        Real = self._real_settings_model()
+        Real.objects.all().delete()
+        Real.objects.create(vc_member_name_pattern=pattern)
+
     def _patch_settings(self, settings_obj):
-        """Patch the deferred import of LibreNMSSettings inside the function."""
+        """Patch the deferred LibreNMSSettings lookup for values the real CharField cannot hold."""
         return patch(
             "netbox_librenms_plugin.models.LibreNMSSettings.objects",
             **{"order_by.return_value.first.return_value": settings_obj},
         )
 
     def test_returns_valid_pattern(self):
-        settings = MagicMock()
-        settings.vc_member_name_pattern = "-SW{position}"
-        with self._patch_settings(settings):
-            assert self._call() == "-SW{position}"
+        """A configured non-empty pattern is read back verbatim from the real settings row."""
+        self._store_pattern("-SW{position}")
+        assert self._call_real() == "-SW{position}"
+
+    def test_returns_default_for_empty_string(self):
+        """An empty-string pattern in the real settings row falls back to the default."""
+        self._store_pattern("")
+        assert self._call_real() == self.DEFAULT
+
+    def test_returns_default_for_whitespace_only(self):
+        """A whitespace-only pattern in the real settings row falls back to the default."""
+        self._store_pattern("   ")
+        assert self._call_real() == self.DEFAULT
+
+    def test_returns_default_when_no_settings(self):
+        """With no settings row persisted, the loader falls back to the default."""
+        self._real_settings_model().objects.all().delete()
+        assert self._call_real() == self.DEFAULT
 
     def test_returns_default_for_none_pattern(self):
+        """A NULL pattern (unreachable via the NOT NULL CharField) falls back to the default."""
         settings = MagicMock()
         settings.vc_member_name_pattern = None
         with self._patch_settings(settings):
             assert self._call() == self.DEFAULT
 
-    def test_returns_default_for_empty_string(self):
-        settings = MagicMock()
-        settings.vc_member_name_pattern = ""
-        with self._patch_settings(settings):
-            assert self._call() == self.DEFAULT
-
-    def test_returns_default_for_whitespace_only(self):
-        settings = MagicMock()
-        settings.vc_member_name_pattern = "   "
-        with self._patch_settings(settings):
-            assert self._call() == self.DEFAULT
-
     def test_returns_default_for_boolean(self):
+        """A boolean pattern (the CharField would stringify it) falls back to the default."""
         settings = MagicMock()
         settings.vc_member_name_pattern = True
         with self._patch_settings(settings):
             assert self._call() == self.DEFAULT
 
-    def test_returns_default_when_no_settings(self):
-        with self._patch_settings(None):
-            assert self._call() == self.DEFAULT
-
     def test_returns_default_on_exception(self):
+        """A DB error while loading settings falls back to the default."""
         with patch(
             "netbox_librenms_plugin.models.LibreNMSSettings.objects",
         ) as mock_objs:
